+++ Sandswept/SandsweptTemporaryEffects.cs
@@ -1,14 +1,21 @@
 using System;
 using Sandswept.Equipment.Standard;
+using Sandswept.Items.Greens;
+using static Sandswept.Items.Greens.MakeshiftPlate;
 
-namespace Sandswept {
-    public class SandsweptTemporaryEffects : MonoBehaviour {
-        public TemporaryVisualEffect GalvanicShield;
-        public void UpdateTemporaryEffects(CharacterBody body) {
-            // body.UpdateSingleTemporaryVisualEffect(ref GalvanicShield, GalvanicCellShield.ShieldEffect, body.bestFitActualRadius * 1.5f, body.HasBuff(Buffs.ParryBuff.instance.BuffDef), "Pelvis");
+namespace Sandswept
+{
+    public class SandsweptTemporaryEffects : MonoBehaviour
+    {
+        public TemporaryVisualEffect makeshiftPlate;
+
+        public void UpdateTemporaryEffects(CharacterBody body)
+        {
+            body.UpdateSingleTemporaryVisualEffect(ref makeshiftPlate, platingOverlayPrefab, MathHelpers.BestBestFitRadius(body), body.GetComponent<PlatingManager>() && body.GetComponent<PlatingManager>().CurrentPlating > 0 && body.inventory && body.inventory.GetItemCount(MakeshiftPlate.instance.ItemDef) > 0);
         }
 
-        internal static void ApplyHooks() {
+        internal static void ApplyHooks()
+        {
             On.RoR2.CharacterBody.UpdateAllTemporaryVisualEffects += OnUpdateVisualEffects;
         }