+++ Sandswept/Survivors/Ranger/SkillDefs/Primary/DirectCurrent.cs
@@ -6,7 +6,7 @@ namespace Sandswept.Survivors.Ranger.SkillDefs.Primary
     {
         public override string Name => "Direct Current";
 
-        public override string Description => "Fire an arcing electric current for $sd150% damage$se. Hitting enemies generates $rc1 Charge$ec, increased up to $rc3$ec on direct hits. <style=cIsUtility>Damage and blast radius ramp up over distance.</style>".AutoFormat();
+        public override string Description => "$suSurging$se. Fire an arcing electric current for $sd150% damage$se. Hitting enemies generates $rc1 Charge$ec, increased up to $rc3$ec on direct hits.".AutoFormat();
 
         public override Type ActivationStateType => typeof(States.Primary.DirectCurrent);
 
@@ -18,7 +18,7 @@ namespace Sandswept.Survivors.Ranger.SkillDefs.Primary
         public override int StockToConsume => 0;
         public override InterruptPriority InterruptPriority => InterruptPriority.Any;
 
-        public override string[] Keywords => [Utils.Keywords.OverdriveFormPrimary];
+        public override string[] Keywords => [Utils.Keywords.Surging, Utils.Keywords.OverdriveFormPrimary];
         public override float GetProcCoefficientData() => 1f;
     }
 }